Update PR template to mention signing
In the past two days we noticed two community PRs that were not signed. We did
not notice, until the PRs were about to be merged, that the merging was blocked.
We should inform the users as early as possible about this requirement to avoid
friction, minimize the risk of losing valuable contributions, and avoid them
having to remake the PR (in the worst case).